hookUpdate: grid's up, state needs to be managed for clicked squares
