+++ src/assets/svgAssets.tsx
@@ -1,0 +1,3 @@
+export const EMPTY_SQUARE = "../assets/iconmonstr-square-4.svg";
+export const FILLED_SQUARE = "../assets/iconmonstr-square-1.svg";
+export const FILLED_X = "../assets/iconmonstr-x-mark-1.svg";
